Abstract servers creation
