+++ applications/archivist/src/gateway/router.ts
