+++ applications/archivist/src/main.ts
@@ -1,27 +1,15 @@
-import {createHTTPServer, type HTTPRouter, logConnections} from '@packages/rpc'
-
-import {archivistRouter} from '@app/router'
+import {setupServer} from '@app/gateway/server'
 import {connectDuct} from '@app/services/duct.service'
 
 const main = async () => {
   await connectDuct()
 
-  const port = process.env.PORT
+  const port = process.env.PORT && parseInt(process.env.PORT, 10)
   if (!port) {
     throw new Error('PORT is not defined')
   }
 
-  const server = createHTTPServer({
-    router: archivistRouter as HTTPRouter,
-  })
-
-  server.listen(port)
-
-  logConnections(server)
-
-  server.on('listening', () => {
-    console.log(`✅ Server listening on ${port}`)
-  })
+  setupServer(port)
 }
 
 main()
