+++ applications/passport/src/gateway/router.ts
