+++ applications/telegraph/src/gateway/router.ts
