+++ packages/rpc/src/index.ts
@@ -1,9 +1,6 @@
-export * from './library/http-server'
-export * from './library/ws-server'
-export * from './library/context'
+export * from './gateways/http-server'
+export * from './gateways/ws-server'
 export * from './utilites/set-cookies'
 export * from './utilites/set-cookies-token'
 export * from './utilites/log-connections'
-export * from './trpc'
-export * from './procudures/token.procedure'
-export * from './procudures/protected.procedure'
+export * from './library/trpc'
