+++ packages/rpc/src/utilites/log-connections.ts
@@ -1,31 +1,13 @@
-import {Server as HTTPServer} from 'node:http'
+import {ServerInterface} from '../gateways/server.interface'
 
-import {Server as WSServer} from 'ws'
+const logConnections = (server: ServerInterface) => {
+  server.onConnection(async (ws) => {
+    console.log(`Connection: ++${await server.getConnectionsCount()}`)
 
-const logConnections = (server: HTTPServer | WSServer) => {
-  server.on('connection', (ws) => {
-    getConnection()
-
-    ws.once('close', () => {
-      getConnection()
+    ws.once('close', async () => {
+      console.log(`Connection: --${await server.getConnectionsCount()}`)
     })
   })
-
-  function getConnection() {
-    if ('getConnections' in server) {
-      server.getConnections((err, connections) => {
-        if (err) {
-          console.error(err)
-        } else {
-          console.log(`Connections: (${connections})`)
-        }
-      })
-    }
-
-    if ('clients' in server) {
-      console.log(`Connections: (${server.clients.size})`)
-    }
-  }
 }
 
 export {logConnections}
